# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YPuddin Train Studio",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/models/downloads": {
      "get": {
        "summary": "List Downloads",
        "operationId": "list_downloads_api_models_downloads_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelDownload"
                  },
                  "type": "array",
                  "title": "Response List Downloads Api Models Downloads Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Start Download",
        "operationId": "start_download_api_models_downloads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/downloads/{download_id}/cancel": {
      "post": {
        "summary": "Cancel Download",
        "operationId": "cancel_download_api_models_downloads__download_id__cancel_post",
        "parameters": [
          {
            "name": "download_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Download Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/downloads/{download_id}/retry": {
      "post": {
        "summary": "Retry Download",
        "operationId": "retry_download_api_models_downloads__download_id__retry_post",
        "parameters": [
          {
            "name": "download_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Download Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models/credentials/{provider}": {
      "put": {
        "summary": "Save Credential",
        "operationId": "save_credential_api_models_credentials__provider__put",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "huggingface",
                "modelscope"
              ],
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear Credential",
        "operationId": "clear_credential_api_models_credentials__provider__delete",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "huggingface",
                "modelscope"
              ],
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models/catalog/{catalog_id}/download": {
      "post": {
        "summary": "Download Catalog",
        "operationId": "download_catalog_api_models_catalog__catalog_id__download_post",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Catalog Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogDownloadRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogDownloadRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "const": "huggingface",
            "title": "Provider",
            "default": "huggingface"
          },
          "mirror": {
            "type": "string",
            "enum": [
              "official",
              "hf-mirror"
            ],
            "title": "Mirror",
            "default": "official"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CatalogDownloadRequest"
      },
...
      "CredentialState": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured"
          }
        },
        "type": "object",
        "required": [
          "configured"
        ],
        "title": "CredentialState"
      },
...
      "CredentialUpdate": {
        "properties": {
          "token": {
            "type": "string",
            "format": "password",
            "title": "Token",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token"
        ],
        "title": "CredentialUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelDownload": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "family": {
            "type": "string",
            "title": "Family"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "provider": {
            "type": "string",
            "enum": [
              "huggingface",
              "modelscope"
            ],
            "title": "Provider",
            "default": "huggingface"
          },
          "mirror": {
            "type": "string",
            "enum": [
              "official",
              "hf-mirror"
            ],
            "title": "Mirror",
            "default": "official"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "target_path": {
            "type": "string",
            "title": "Target Path"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "downloading",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "downloaded_bytes": {
            "type": "integer",
            "title": "Downloaded Bytes",
            "default": 0
          },
          "total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Bytes"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "dtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dtype"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "family",
          "kind",
          "source_url",
          "filename",
          "target_path",
          "status",
          "created_at"
        ],
        "title": "ModelDownload"
      },
      "ModelDownloadRequest": {
        "properties": {
          "family": {
            "type": "string",
            "enum": [
              "anima",
              "krea2",
              "tagger"
            ],
            "title": "Family"
          },
          "kind": {
            "type": "string",
            "enum": [
              "dit",
              "text_encoder",
              "vae",
              "tagger"
            ],
            "title": "Kind"
          },
          "provider": {
            "type": "string",
            "enum": [
              "huggingface",
              "modelscope"
            ],
            "title": "Provider",
            "default": "huggingface"
          },
          "mirror": {
            "type": "string",
            "enum": [
              "official",
              "hf-mirror"
            ],
            "title": "Mirror",
            "default": "official"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "repo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Id"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "dtype": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "bf16",
                  "fp16",
                  "fp32",
                  "fp8"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Dtype"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "family",
          "kind"
        ],
        "title": "ModelDownloadRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}